requirements.txt
defusedxml==0.5.0
Django==2.1.7
djangorestframework==3.9.2
mysqlclient==1.4.2.post1
oauthlib==3.0.1
Pillow==5.4.1
pkg-resources==0.0.0
python3-openid==3.1.0
pytz==2018.9
requests-oauthlib==1.2.0
six==1.12.0
